Add theme, stuff

+++ run.sh
@@ -1,4 +1,6 @@
 #!/bin/sh
 . ./bin/activate
 ./bin/pelican content -s pelicanconf.py -t themes/pelican-elegant-1.3
-rsync -razvP --delete-after output/ cmr@octayn.net:twir
+if [ "$REALLY_DEPLOY" = "1" ]; then
+    rsync -razvP --delete-after output/ cmr@octayn.net:twir
+fi
